transcendence/requirements.txt
asgiref==3.8.1
astroid==3.3.8
autopep8==2.0.4
black==24.10.0
click==8.1.8
dill==0.3.9
Django==5.1.4
django-cors-headers==4.0.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docstring-to-markdown==0.15
flake8==7.1.1
isort==5.13.2
jedi==0.19.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.10.1
pylint==3.3.3
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
pytoolconfig==1.3.1
rope==1.13.0
ruff==0.8.6
snowballstemmer==2.2.0
sqlparse==0.5.3
tomlkit==0.13.2
ujson==5.10.0
whatthepatch==1.0.7
yapf==0.43.0
